% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bintanja}
\alias{bintanja}
\title{Dataset: Bintanja's Global sea level curve (0-3000000 BP)}
\format{
A vector with:
\describe{
 \item{year_before_after_present}{years before or after present}
 \item{sea_level_m}{meters below or above the reference sea level expressed in m.}

}
}
\source{
Bintanja, R., van de Wal, R. (2008) North American ice-sheet dynamics and the onset of 100,000-year glacial cycles. Nature 454, 869–872 (2008). \doi{https://doi.org/10.1038/nature07158}
}
\usage{
bintanja
}
\description{
Global sea level curve of Bintanja & van de Wal (2008) from 0 to 3000000 BP.
}
\examples{
curve <- bintanja

}
\keyword{datasets}
