% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{codebook}
\alias{codebook}
\title{Codebook}
\format{
A data frame with 2 variables:
\describe{
 \item{code}{Numerical code}
 \item{value}{Descriptive value/label}
}
}
\source{
Texas Department of Licensing and Regulation
}
\usage{
codebook
}
\description{
Table for decoding numerical codes to descriptive values from TDLR
}
\keyword{datasets}
