% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabs_classes.R
\name{create_tabs_class}
\alias{create_tabs_class}
\title{Create an Object of Class 'tabs'}
\usage{
create_tabs_class(x)
}
\arguments{
\item{x}{output of reconstruct}
}
\value{
An object of class 'tabs'.
}
\description{
Constructs an S3 object of class 'tabs' containing reconstruct output.
}
\keyword{internal}
