% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cutler}
\alias{cutler}
\title{Dataset: Cutler's Global sea level curve (0-140000 BP)}
\format{
A vector with:
\describe{
 \item{year_before_after_present}{years before or after present}
 \item{sea_level_m}{meters below or above the reference sea level expressed in m.}

}
}
\source{
Cutler, Kirsten B; Edwards, Ross L; Taylor, Frederick W; Cheng, H; Adkins, Jess F; Gallup, Christina D; Cutler, P M; Burr, George S; Bloom, Arthur L (2003): Rapid sea-level fall and deep-ocean temperature change since the last interglacial period. Earth and Planetary Science Letters, 206(3-4), 253-271. \doi{https://doi.org/10.1016/S0012-821X(02)01107-X}
}
\usage{
cutler
}
\description{
Global sea level curve of Cutler et al. (2003) from 0 to 140000 BP.
}
\examples{
curve <- cutler

}
\keyword{datasets}
