% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{features}
\alias{features}
\title{Dataset: features}
\format{
data frame 
\describe{
 \item{feature_class}{feature class}
 \item{feature_class_description}{description of the feature class}
 \item{feature}{feature names}

}
}
\source{
GeoNames \url{https://download.geonames.org/export/dump/}
}
\usage{
features
}
\description{
Features that can be defined as labeling points in the default dataset. point_reference dataset has been retrieved from geoNames.
}
\examples{
f <- features

}
\keyword{datasets}
