% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{funza}
\alias{funza}
\title{Dataset: Funza (29500 up to 1000000 BP)}
\format{
A vector with:
\describe{
 \item{year_before_after_present}{years before or after present}
 \item{tree_line_m}{tree line expressed in meters above the present day reference sea level.}

}
}
\source{
Flantua et al. (2019) \doi{https://doi.org/10.3389/fevo.2021.615223}
}
\usage{
funza
}
\description{
The ‘páramos’ UFL, i.e. high altitude mountain ecosystem in Venezuela (Northern Andes), during the last 1 million years, regularized to intervals of 1 ky.
}
\examples{
curve <- funza

}
\keyword{datasets}
