% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_reclabs.R
\name{get_reclabs}
\alias{get_reclabs}
\title{get_reclabs}
\usage{
get_reclabs(labs, recvect, noise, verbose)
}
\arguments{
\item{labs}{labeling dataset}

\item{recvect}{polygon dataset}

\item{noise}{number of cells considered as noise}

\item{verbose}{boolean. FALSE: No messages are printed. TRUE: Standard verbose mode, providing progress bar. 2: Very verbose mode, displaying detailed information.}
}
\value{
SpatVector
}
\description{
reconstruct the labels based on labeling dataset. In a second step polygons are labelled per time step based on their most recent appearance
}
\author{
Johannes De Groeve
}
\keyword{internal}
