% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{IPCC_global_mean}
\alias{IPCC_global_mean}
\title{Dataset: IPCC future predictions (2021-2100)}
\format{
A list four RSL vectors
\describe{
 \item{year_before_after_present}{years before or after present}
 \item{sea_level_m}{meters below or above the reference sea level expressed in m.}

}
}
\source{
IPCC, Kirsten B; Edwards, Ross L; Taylor, Frederick W; Cheng, H; Adkins, Jess F; Gallup, Christina D; Cutler, P M; Burr, George S; Bloom, Arthur L (2003): Rapid sea-level fall and deep-ocean temperature change since the last interglacial period. Earth and Planetary Science Letters, 206(3-4), 253-271. \doi{https://doi.org/10.1016/S0012-821X(02)01107-X} \url{https://interactive-atlas.ipcc.ch/regional-information}
}
\usage{
IPCC_global_mean
}
\description{
Mean global sea level rise according to different scenarios (ssp1,ssp2,ssp3,ssp5) for 2021-2040, 2041-2060 and 2081-2100. The global means were calculated from global raster datasets available for download in the interactive IPCC atlas.
}
\examples{
curves <- IPCC_global_mean
 
}
\keyword{datasets}
