% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lambeck}
\alias{lambeck}
\title{Dataset: Lambeck's Global sea level curve (0-35000 BP)}
\format{
A vector with:
\describe{
 \item{year_before_after_present}{years before or after present}
 \item{sea_level_m}{meters below or above the reference sea level expressed in m.}

}
}
\source{
Lambeck, Kurt; Rouby, Hélène; Purcell, Anthony; Sun, Y; Sambridge, Malcom (2014): Sea level and global ice volumes from the Last Glacial Maximum to the Holocene. Proceedings of the National Academy of Sciences, 111(43), 15296-15303. \doi{https://doi.org/10.1073/pnas.1411762111}
}
\usage{
lambeck
}
\description{
Global sea level curve of Lambeck et al. (2014) from 0 to 35000 BP.
}
\examples{
curve <- lambeck

}
\keyword{datasets}
