% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{regions}
\alias{regions}
\title{Dataset: regions}
\format{
vector
\describe{
 \item{dataset}{the dataset (islands,mountains) for which the region definition can be used}
 \item{region}{region type}
 \item{name}{name of the region}
 \item{name_ascii}{name in ascii format}
}
}
\source{
Islands: Sayre et al. 2019 \doi{https://doi.org/10.1080/1755876X.2018.1529714}

Plates: Bird, P. (2003) \doi{https://doi.org/10.1029/2001GC000252}

Archipelagoes: Weigelt et al. 2013 \doi{https://doi.org/10.5061/dryad.fv94v}

Countries: from rnaturalearth \url{https://docs.ropensci.org/rnaturalearth/}

Mountains: Snethlage et al. 2022 \doi{https://doi.org/10.48601/earthenv-t9k2-1407}
}
\usage{
regions
}
\description{
mountain ranges at different hierarchical levels, islands, archipelagoes, countries and plates that can be used for region selection
}
\examples{
r <- regions

}
\keyword{datasets}
