% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{setup}
\alias{setup}
\title{Setup - Download of default datasets}
\usage{
setup()
}
\value{
No return value, called for side effects
}
\description{
download default datasets (labs, topo, curve) in default or custom directory. The datasets require 15 GB disk space.
}
\examples{

if(interactive()){

# download labs, topo and curve and store in custom or default directory
setup()

# reset setup
# if you want the install the datasets under a different path
options(tabs.datasetPath=NULL)
setup()
}

}
\author{
Johannes De Groeve
}
