% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampledata.R
\name{sporades}
\alias{sporades}
\title{Sporades sample dataset including topo, labs, curve and correction}
\usage{
sporades()
}
\value{
List of input sample datasets (topo, labs, curve, correction)
}
\description{
Sample dataset for the Sporades Archipelago in Greece
}
\author{
Johannes De Groeve
}
