% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetProjects.R
\name{GetProjects}
\alias{GetProjects}
\title{Get registered projects}
\usage{
GetProjects(
  county = NULL,
  city = NULL,
  reg_begin = "",
  reg_end = "",
  owner = "",
  project = "",
  facility = "",
  address = ""
)
}
\arguments{
\item{county}{The Texas county where the physical project is}

\item{city}{The Texas city where the physical project is}

\item{reg_begin}{The start date for searching projects Date Format: mm/dd/yy}

\item{reg_end}{The cutoff date for searching projects Date Format: mm/dd/yy}

\item{owner}{Search by the owner of a project}

\item{project}{Search by project name}

\item{facility}{Search by the facility name}

\item{address}{Search by the physical address of a project}
}
\value{
Projects based on the given criteria
}
\description{
Returns a df of projects registered in the TABS database
}
\examples{
GetProjects(
  county = "Caldwell",
  reg_begin = "03/01/24",
  reg_end = "04/01/24"
)
GetProjects(
  city = "Austin",
  owner = "Tesla",
  reg_begin = "01/01/20",
  reg_end = "01/01/24"
)
}
