% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TABSdecoder.R
\name{TABSdecoder}
\alias{TABSdecoder}
\title{Recode values to and from TDLR}
\usage{
TABSdecoder(codes, ..., codebook_df = dat)
}
\description{
Will submit appropriate codes to TDLR and return plain language values to user
}
